// frontend/src/generated/types/commonResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { CommonResponseData } from './commonResponseData';

export interface CommonResponse {
  success?: boolean;
  message?: string;
  error?: string;
  data?: CommonResponseData;
}